requirements.txt
Flask~=3.1.0
Flask-Cors
pywidevine~=1.8.0
pyplayready~=0.6.0
requests~=2.32.3
protobuf~=4.25.6
PyYAML~=6.0.2
mysql-connector-python
bcrypt